# Add expansion support to HHashTable to optimize HashAgg