With reference to https://github.com/greenplum-db/gpdb/pull/2155, this PR adds support for an expandable hash table for HashAgg. 

* When creating the hash table, instead of determining the number of buckets (nbuckets) only by the available memory, it now computes the nbuckets as a function of estimated number of entries by the planner/optimizer. So in the extreme case of the example below it reduces the iteration time from 419.731 ms to 31 ms, which is a 10X improvement:
```
template1=# explain analyze select i, sum(j) from foo group by i;explain analyze select i, sum(j) from foo
 group by i;
                                     QUERY PLAN
-------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=992.79..606.58 rows=1000 width=12)
   Rows out:  0 rows at destination with 31 ms to end, start offset by 0.355 ms.
   ->  HashAggregate  (cost=992.79..606.58 rows=334 width=12)
         Group By: i
         Rows out:  0 rows (seg0) with 0.210 ms to end, start offset by 21 ms.
         Executor memory:  16K bytes avg, 16K bytes max (seg0).
         ->  Seq Scan on foo  (cost=0.00..961.00 rows=28700 width=8)
               Rows out:  0 rows (seg0) with 0.131 ms to end, start offset by 21 ms.
 Slice statistics:
   (slice0)    Executor memory: 322K bytes.
   (slice1)    Executor memory: 102K bytes avg x 3 workers, 102K bytes max (seg0).
 Statement statistics:
   Memory used: 8388608K bytes
 Optimizer status: legacy query optimizer
 Total runtime: 31.828 ms
(15 rows)
```
* When reloading spilled groups a spill file, we can re-compute the number of buckets based on the number of entries that were written to the file. This means in the reloading case, in most cases we will not need to expand the hash table - and in the best case reduce the allocation and zero-ing costs.
* This PR also includes some refactoring code that fixes the memory computations. This should mean we stay within memory bounds more often, although many corner cases still remain. See the FIXME comment in `spill_hash_table()`. I have a created a separate story in the QX backlog to tackle that.
* I also fixed the EXPLAIN ANALYZE metrics in the context of an expanding hashtable (total_buckets & chainlength).
* The hash table can spill in multiple "layers" i.e spill again when it is processing an already spilled file. There were no tests in ICG testing the correctness for level > 1, so I added some tests for that. Any other recommendation for testing the expansion is welcome - although I suspect a component will tested in multiple places.