@Eulerizeit 2017-04-14
> in the extreme case of the example below it reduces the iteration time from 419.731 ms to 31 ms, which is a 10X improvement

Nice!!!